requirements.txt
asgiref==3.4.0
backports.entry-points-selectable==1.1.0
distlib==0.3.2
Django==3.2.5
django-cors-headers==3.8.0
django-crispy-forms==1.12.0
django-filter==2.4.0
djangorestframework==3.12.4
filelock==3.0.12
gunicorn==20.1.0
Markdown==3.3.4
MouseInfo==0.1.3
numpy==1.21.2
Pillow==8.3.1
platformdirs==2.2.0
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
Pygments==2.10.0
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.27
PyTweening==1.0.3
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
virtualenv==20.7.2
whitenoise==5.3.0
